Remove overwrite of #new on generic_files_controller.
Unfortunately can't really make sense of the commit message or comment
from when this was added. Looks to have been related to
https://github.com/projecthydra/sufia/commit/d852947ba50876177a96e3f2b86d0b6714b28269
but the logic of that commit seems fine. Cannot find a relevant bug
report or issue. todo list from that day mentions an upload error. there
have been many of those. uploads seem fine with this change.