pf phrase boost same fields as qf
I don't see why not to start this way. pf means if the words they entered
appear _near_ each other in the field, it'll be a higher boost. Seems like
we'd want this in all the same fields

Theoretically we might want to boost the pf even HIGHER than the qf,
but may or may not be neccesary and we don't have a good way to keep
that organized right now.