override template to prevent empty row for parent object relationship when empty (In Generic Work or Part of)
The architecture here is very convoluted and odd, this is the least intrusive and least
likely to break in upgrades way I found to do it. I don't think anything is added by adding
another layer of indirection and moving the empty-skip stuff to configuration, it's too complex
as it is.