avoid expensive application_name call
It turns out calling Sufia helper  is quite expensive, too
expensive to do twice per member on a page that could have hundreds of members.
Sufia::SufiaHelperBehavior#application_name

We could have tried to figure out why it was expensive and optimized it, but
we don't need it enough to be worth the time at present.