Avoid calling representative_presenter for each child work on a parent show page
It results in three trips to Solr and is too expensive.

To avoid it, we index the data we need directly on the work itself.

This does seem to significantly improve performance, and is probably the cause of
our last remaining perf problem crosses fingers.

I separately PR'd the indexing changes to master. I did them here too, which
might cause a conflict on merge, but since the GenericWorkIndexer changes
paths between the branches, I was worried it might end up lost entirely
and not properly merged without a conflcit, I'd rather a conflict so
we make sure it winds up here.