Human-friendlier download file names in 'save as'
We wanted to include the 'index' number (page X of total) in the name, for
even more friendliness and sorting right locally, but was too much of a pain
to get and track that with current architecture.

This architecture is def getting a bit sprawling, those :small :medium :large
magic key names are a bit non-ideal, but it's what we've got for now.