rights statement back on viewer download menu
a bit hacky, leaving this in the DOM as HTML for JS to get. But it works.

member_helper and image_service_helper are getting expansive, would be good
to refactor into something other than a mess of helper methods, some kind
of OO. But tricky and good enough for now.