one cheesy way to temporarily prevent staff login and authentication-required actions
Uses a devise feature not quite intended for this.

The UX isn't great.

If you try to login, you will get a login screen, but logging in succesfully there will be bounce you back to login screen again, this time with the message 'Sorry, logins are temporarily disabled for some software maintenance.'

If you were already logged in, and try to access the app, you'll be bounced to the login screen with that message too. Even if what you were doing did not require a login.

I'll see if I can make it a little bit better.