move blacklight_oai_provider to it's own controller
Gives us a way to modify the search builder so we can exclude collections. There might be some other way to do this that I couldn't figure out that would be more preferred by blacklight_oai_provider, not sure

This might be a clear way to have the code anyway? Not sure. But it works.