Complete override of CreateWithFilesActors and it's dependencies
It now does the work of CreateWithRemoteFilesActor too (that one has been zeroed out).

To avoid doing things in the foreground web server loop that should be in a bg job, and generally make this saner and easier to reason about. It was such a mess, it just made sense to take it over.