on ingest, create FileSets serially
in a new job. cause file sets can't be attached to work in parallel, they were fighting with each other.

but we still need it in the bg not fg cause it's slow on bulk ingest.

Then the new AttachFileSetsJobs kicks off a job per file to actually ingest bytestreams