upgrade active_fedora to 11.5.2
From 11.1.6

To upgrade activefedora we also had to upgrade other dependencies, so as to be compatible.

This is to get a version of active-fedora which fixes this bug:
https://github.com/samvera/hyrax/issues/2108

Which seems to be responsible for this bad behavior in our app:
https://github.com/sciencehistory/chf-sufia/issues/1105

It's a mystery why we didn't run into this active-fedora bug until now, but this does seem to fix it.