Restore member_audio_tag presentation of actual audio players in the 'listing' for staff members only.
Accidentally removed this when I did "Remove special handling for audio files in 'downloads' menu in standard file/thumb component".

We don't really need/want this feature, but didn't mean to remove it. And don't want to think through whether removing it still gives admins the UI they need to edit things, since this one was already at least hypothetically approved.

Fixed the specs to not expect the DOWNLOAD links that we did intentionally remove.

This reverts commit 26b23ce818769a9015329ed13b168d6e25873b9f.